+++ src/Catalog/Product/ProductBridge.php
@@ -65,23 +65,27 @@ class ProductBridge
             ->addFilter(new ProductAvailableFilter($context->getSalesChannel()->getId()))
             ->addFilter(new EqualsFilter('product.parentId', null))
             ->addAssociations([
-                'media',
-                'properties',
-                'properties.group',
-                'manufacturer',
-                'options',
-                'seoUrls',
                 'categories',
-                'visibilities',
-                'variation',
+                'manufacturer',
+                'media',
+                'options',
                 'prices',
                 'prices.rule',
+                'properties',
+                'properties.group',
+                'seoUrls',
+                'visibilities',
+                'variation',
 
                 'children',
+                'children.categories',
+                'children.manufacturer',
                 'children.media',
+                'children.options',
+                'children.prices',
+                'children.prices.rule',
                 'children.properties',
                 'children.properties.group',
-                'children.options',
                 'children.seoUrls',
                 'children.unit'
             ])