+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# 0.0.13
+- added variant (child) manufacturer, price & property export
+- fixed issue with missing payment/shipping method notifications showing for customer after a mobile order is made
+
 # 0.0.12
 - added exclusion of shopgate free shipping method from check_cart
-- fixed base_price currency formatting & adjusted overall presentation
 - added customer shipping/billing addresses to check_cart for rule calculations
+- fixed base_price currency formatting & adjusted overall presentation
 
 # 0.0.11
 - added shopgate free shipping method for 0 priced imports